Reduce the number of CI runs
The build requirements for the Sphinx environment should only ever change
when the conf.py or custom_conf.py changes. However, the current
workflow runs it for every push or pull request. Change it to only run
when the dependencies can change.

Note: there is some risk in that an upstream dependency may change under
the covers and cause future sphinx venv failures, however building the
sphinx environmet takes 30+ minutes. The tradeoff of catching it late is
reasonable vs spending 30 minutes each time a push or PR occurs.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>