docs: Fix documentation note tags and other issues
The Canonical OpenStack documentation pages contain quite a few instances
in which note tags and similar are supposed to be used, however, the
synthax is not RST-native, and the tags themselves are not rendered (the
documentation pages simply contains that text).

This addresses the issue.

Additionally, fixes a few minor issues.