Temporarily suspend PDF builds
PDF Builds are hanging on read the docs and needs to be further
debugged and diagnosed. This is causing the RTD builds to timeout
and fail, which is preventing updates from being published.

This change temporarily removes the PDF builds for the moment to
allow an update of docs while the PDF builds are debugged.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>